move upload to models
