+++ routes/api/users.js
@@ -7,7 +7,7 @@ const passport = require('passport');
 const router = express.Router();
 const validateRegisterInput = require('../../validation/register');
 const validateLoginInput = require('../../validation/login');
-const upload = require('../../middleware/upload');
+const upload = require('../../models/Upload');
 
 // router.get("/test", (req, res) => res.json({ msg: "This is the users route" }));
